% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-AgentSeurat.R
\name{get_decisions}
\alias{get_decisions}
\title{Return the accumulated decision log}
\usage{
get_decisions(obj)
}
\arguments{
\item{obj}{An AgentSeurat object.}
}
\value{
A list of decision records.
}
\description{
Return the accumulated decision log
}
