% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-checkpoint.R
\name{load_checkpoint}
\alias{load_checkpoint}
\title{Load an AgentSeurat checkpoint}
\usage{
load_checkpoint(path)
}
\arguments{
\item{path}{Path to a `.qs2` file saved by [save_checkpoint()].}
}
\value{
The loaded AgentSeurat object.
}
\description{
Load an AgentSeurat checkpoint
}
