% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{chat_claude}
\alias{chat_openai}
\alias{chat_deepseek}
\alias{chat_grok}
\alias{chat_qwen}
\alias{chat_kimi}
\alias{list_chat_providers}
\alias{make_chat_fn_claude}
\alias{make_chat_fn_openai_compat}
\alias{make_chat_fn_openai_compatible}
\alias{save_checkpoint}
\alias{load_checkpoint}
\alias{export_decisions}
\alias{get_decisions}
\alias{get_script}
\alias{get_figures}
\alias{get_token_usage}
\alias{upgrade_checkpoint}
\alias{token_usage_summary}
\alias{token_usage_reset}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{agentomicsCore}{\code{\link[agentomicsCore:chat_claude]{chat_claude()}}, \code{\link[agentomicsCore:chat_deepseek]{chat_deepseek()}}, \code{\link[agentomicsCore:chat_grok]{chat_grok()}}, \code{\link[agentomicsCore:chat_kimi]{chat_kimi()}}, \code{\link[agentomicsCore:chat_openai]{chat_openai()}}, \code{\link[agentomicsCore:chat_qwen]{chat_qwen()}}, \code{\link[agentomicsCore:export_decisions]{export_decisions()}}, \code{\link[agentomicsCore:get_decisions]{get_decisions()}}, \code{\link[agentomicsCore:get_figures]{get_figures()}}, \code{\link[agentomicsCore:get_script]{get_script()}}, \code{\link[agentomicsCore:get_token_usage]{get_token_usage()}}, \code{\link[agentomicsCore:list_chat_providers]{list_chat_providers()}}, \code{\link[agentomicsCore:load_checkpoint]{load_checkpoint()}}, \code{\link[agentomicsCore:make_chat_fn_claude]{make_chat_fn_claude()}}, \code{\link[agentomicsCore:make_chat_fn_openai_compat]{make_chat_fn_openai_compat()}}, \code{\link[agentomicsCore:make_chat_fn_openai_compatible]{make_chat_fn_openai_compatible()}}, \code{\link[agentomicsCore:save_checkpoint]{save_checkpoint()}}, \code{\link[agentomicsCore:token_usage_reset]{token_usage_reset()}}, \code{\link[agentomicsCore:token_usage_summary]{token_usage_summary()}}, \code{\link[agentomicsCore:upgrade_checkpoint]{upgrade_checkpoint()}}}
}}

