% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-AgentSeurat.R
\name{get_seurat}
\alias{get_seurat}
\title{Extract the underlying Seurat object (or list)}
\usage{
get_seurat(obj)
}
\arguments{
\item{obj}{An AgentSeurat object.}
}
\value{
A Seurat object or a list of Seurat objects.
}
\description{
Extract the underlying Seurat object (or list)
}
